Some live FGH database test utils that save time
